20230212 회의
controller 작성중
- 관리자 페이지용 remove/modify
- DB 수정 및 추가 예정

+++ PickaBoo/jam/src/main/java/com/funny/jam/Entity/RawCrawlingEntity.java
@@ -15,11 +15,10 @@ import lombok.Setter;
 @NoArgsConstructor
 @Getter
 @Setter
-public class RawCrawlingEntity 
-{
+public class RawCrawlingEntity {
     @Id
-    private double seq;
+    private Long seq;
     private String siteLinkInfo;
     private String imageRemotPathInfo;
-    private String keywordInfo; 
+    private String keywordInfo;
 }
